push_notifications_0.51
notifications are now their own asset
they are also relevant to the newest item added
