+++ views/new_fooditem.pug
@@ -103,6 +103,7 @@ block container
 	script
 		include assets/upload.js
 		include assets/add.js
+		include assets/push_notifications.js
 		include assets/lib/quagga.js
 		include assets/barcode.js
 		include assets/js-cookie.js
